(Engineering) Fix compiler warnings issued by (Apple) clang
Changed type of loop control variable to int to remove warning of comparison
of inconsistent integer types (int vs size_t).

Issue #444

+++ src/Rose/Logger.C
@@ -52,11 +52,11 @@ void Logger::display(Level const lvl, std::string const & message) {
   } else {
     bool indentable = indent.size() > 0 && depth > 0;
     if (indent_first && indentable)
-      for (size_t i = 0; i < depth; i++)
+      for (int i = 0; i < depth; i++)
         std::cerr << indent;
     std::cerr << level_to_string(lvl);
     if (!indent_first && indentable)
-      for (size_t i = 0; i < depth; i++)
+      for (int i = 0; i < depth; i++)
         std::cerr << indent;
     std::cerr << message << std::endl;
   }
